fix z3 in a slightly nicer way
the problem wasn't off-by-one arithemetic: it's that if you allow floor
division, there are multiple valid values for lvar which will produce
the correct target_var, and z3 just happened to be picking the wrong one
in some cases. If you limit the model to only allowing true division,
then there's a unique solution.